finish norming
