+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: jteissie <jteissie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/25 13:43:42 by bthomas           #+#    #+#             */
-/*   Updated: 2024/07/21 18:27:51 by jteissie         ###   ########.fr       */
+/*   Updated: 2024/07/21 18:37:13 by jteissie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,28 +31,11 @@ int	parse_data(t_data *data)
 
 int	tokenize_data(t_data *data)
 {
-	t_token	*roaming;
-
 	if (valid_input(data->input, data))
 	{
 		if (lexer(data))
 		{
-			if (data->token)
-				roaming = data->token;
-			else
-				roaming = NULL;
-			while (roaming && roaming->next)
-				roaming = roaming->next;
-			ft_putstr_fd("minishell: syntax error near unexpected token ", 2);
-			data->errcode = 2;
-			if (roaming)
-			{
-				ft_putchar_fd('\'', 2);
-				ft_putstr_fd(roaming->lexstr, 2);
-				ft_putchar_fd('\'', 2);
-			}
-			ft_putchar_fd('\n', 2);
-			free_lexmem(data);
+			throw_lexer_error(data);
 			return (PANIC);
 		}
 	}
